\name{ReturnSeries}
\alias{ReturnSeries}
\docType{data}
\title{
Return Series Data
}
\description{
Monthly return series of International Soyabean oil starting from January 1980
}
\usage{data("ReturnSeries")}
\format{
  A data frame with 86 observations on the following variable.
  \describe{
    \item{\code{return}}{a numeric vector}
  }
}
\details{
Dataset contain 86 0bservations of monthly return series of International soyabean price. It is obtained from World Bank "Pink sheet"
}
\source{
https://www.worldbank.org/en/research/commodity-markets
}
\references{
https://www.worldbank.org/en/research/commodity-markets
}
\examples{
data(ReturnSeries)
}
\keyword{datasets}
